reinstantiate profile page
