+++ app/blueprints/multilingual/templates/multilingual/base.html
@@ -47,6 +47,9 @@
                 <li class="nav-item">
                     <a href = "{{ url_for('multilingual.count_logins') }}" class = "nav-link">{{_('News')}}</a>
                 </li>
+                <li class="nav-item">
+                    <a href = "{{ url_for('multilingual.profile') }}" class = "nav-link">{{_('My Profile')}}</a>
+                </li>
 
                 <li class="nav-item">
                     <a href = "{{ url_for('multilingual.contact') }}" class = "nav-link">{{_('Mail us')}}</a>
