+++ app/blueprints/multilingual/templates/multilingual/profile.html
@@ -371,6 +371,11 @@ window.onclick = function(event) {
 </div>
 {% endif %}
   <!-- Content Column -->
+
+  <h2 class="m-0 font-weight-bold text-primary">Hi, {{username}}</h2>
+
+
+  
 {% if ((phase == 2 and (group == 2 or group == 3)) or (group == 4)) %}
   <div class="row">
     {% endif %}
